requirements.txt
streamlit
pandas
numpy
Pillow
opencv-python
loguru
lmdb
sqlalchemy
psycopg2
xlrd
tqdm
imgaug
tensorflow-cpu
tensorflow-addons
argparse
h5py
PyTurboJPEG